+++ x/vpool/types/snapshot.go
@@ -28,6 +28,11 @@ func (s ReserveSnapshot) Validate() error {
 		return err
 	}
 
+	if (s.BaseAssetReserve.String() == "<nil>") || (s.QuoteAssetReserve.String() == "<nil>") {
+		// prevents panics from usage of 'new(Dec)' or 'sdk.Dec{}'
+		return fmt.Errorf("nil dec value in snapshot. snapshot: %v", s.String())
+	}
+
 	if s.BaseAssetReserve.IsNegative() {
 		return fmt.Errorf("base asset reserve from snapshot cannot be negative: %d", s.BaseAssetReserve)
 	}
@@ -36,8 +41,15 @@ func (s ReserveSnapshot) Validate() error {
 		return fmt.Errorf("quote asset reserve from snapshot cannot be negative: %d", s.QuoteAssetReserve)
 	}
 
-	if s.TimestampMs < 0 {
-		return fmt.Errorf("timestamp from snapshot cannot be negative: %d", s.TimestampMs)
+	// -62135596800000 in Unix milliseconds is equivalent to "0001-01-01 00:00:00 +0000 UTC".
+	// This is the earliest possible value for a ctx.blockTime().UnixMilli()
+	const earliestMs int64 = -62135596800000
+	if s.TimestampMs < earliestMs {
+		snapshotTime := time.Unix(s.TimestampMs/1e3, s.TimestampMs%1e3).UTC()
+		earliestTime := time.Unix(earliestMs/1e3, earliestMs%1e3).UTC()
+		return fmt.Errorf("snapshot time (%s, milli=%v) should not be before "+
+			"earliest possible UTC time (%s, milli=%v): ",
+			snapshotTime, s.TimestampMs, earliestTime, earliestMs)
 	}
 
 	return nil
@@ -54,7 +66,7 @@ func (s ReserveSnapshot) GetLowerMarkPriceFluctuationLimit(fluctuationLimitRatio
 }
 
 // getMarkPrice returns the price of the mark price at the moment of the snapshot.
-// It is the equivalent of getMarkPrice from VPool
+// It is the equivalent of getMarkPrice from Vpool
 func (s ReserveSnapshot) getMarkPrice() sdk.Dec {
 	if s.BaseAssetReserve.IsNil() || s.BaseAssetReserve.IsZero() ||
 		s.QuoteAssetReserve.IsNil() || s.QuoteAssetReserve.IsZero() {